Moved certain size variables to int64.

+++ api/blob_test.go
@@ -37,12 +37,12 @@ func Test_parseValidBlob(t *testing.T) {
 		hdr, err := p.ParseHeader()
 		util.AssertNoErr(t, err)
 		util.Assert(t, hdr.Type() == ObjectBlob)
-		util.Assert(t, hdr.Size() == len(v))
+		util.Assert(t, hdr.Size() == int64(len(v)))
 
 		o, err := p.ParsePayload()
 		util.AssertNoErr(t, err)
 		util.Assert(t, o.Header().Type() == ObjectBlob)
-		util.Assert(t, o.Header().Size() == len(v))
+		util.Assert(t, o.Header().Size() == int64(len(v)))
 		util.Assert(t, o.ObjectId() == nil) // wasn't set in the test scenario
 		util.AssertPanicFree(t, func() {
 			b := o.(*Blob)
@@ -76,7 +76,7 @@ func Test_parseBlobBadSize(t *testing.T) {
 		hdr, e := p.ParseHeader()
 		util.AssertNoErr(t, e)
 		util.Assert(t, hdr.Type() == ObjectBlob)
-		util.Assert(t, hdr.Size() == size)
+		util.Assert(t, hdr.Size() == int64(size))
 
 		// should not be able to parse
 		_, e = p.ParsePayload()
